+++ onmt/train/Trainer.lua
@@ -19,28 +19,66 @@ local function eval(model, data)
   return math.exp(loss / totalWords)
 end
 
-------------------------------------------------------------------------------------------------------------------
+---------------------------------------------------------------------------------
 
 local Trainer = torch.class('Trainer')
 
 local options = {
-  {'-save_every',              0 ,    [[Save intermediate models every this many iterations within an epoch.
-                                            If = 0, will not save models within an epoch. ]],
-                                      {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
-  {'-report_every',            50,    [[Print stats every this many iterations within an epoch.]],
-                                      {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
-  {'-async_parallel',          false, [[Use asynchronous parallelism training.]]},
-  {'-async_parallel_minbatch', 1000,  [[For async parallel computing, minimal number of batches before being parallel.]],
-                                      {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
-  {'-start_iteration',         1,     [[If loading from a checkpoint, the iteration from which to start]],
-                                         {valid=onmt.utils.ExtendedCmdLine.isInt(1)}},
-  {'-end_epoch',               13,    [[The final epoch of the training]],
-                                      {valid=onmt.utils.ExtendedCmdLine.isInt(1)}},
-  {'-start_epoch',             1,     [[If loading from a checkpoint, the epoch from which to start]],
-                                      {valid=onmt.utils.ExtendedCmdLine.isInt(1)}},
-  {'-curriculum',              0,     [[For this many epochs, order the minibatches based on source
-                                            sequence length. Sometimes setting this to 1 will increase convergence speed.]],
-                                      {valid=onmt.utils.ExtendedCmdLine.isUInt()}}
+  {
+    '-save_every', 5000,
+    [[Save intermediate models every this many iterations within an epoch.
+      If = 0, will not save intermediate models.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isUInt()
+    }
+  },
+  {
+    '-report_every', 50,
+    [[Report progress every this many iterations within an epoch.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isUInt()
+    }
+  },
+  {
+    '-async_parallel', false,
+    [[When training on multiple GPUs, update parameters asynchronously.]]
+  },
+  {
+    '-async_parallel_minbatch', 1000,
+    [[In asynchronous training, minimal number of sequential batches before being parallel.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isUInt()
+    }
+  },
+  {
+    '-start_iteration', 1,
+    [[If loading from a checkpoint, the iteration from which to start.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isInt(1)
+    }
+  },
+  {
+    '-start_epoch', 1,
+    [[If loading from a checkpoint, the epoch from which to start.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isInt(1)
+    }
+  },
+  {
+    '-end_epoch', 13,
+    [[The final epoch of the training.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isInt(1)
+    }
+  },
+  {
+    '-curriculum', 0,
+    [[For this many epochs, order the minibatches based on source length.
+      Sometimes setting this to 1 will increase convergence speed.]],
+    {
+      valid = onmt.utils.ExtendedCmdLine.isUInt()
+    }
+  }
 }
 
 function Trainer.declareOpts(cmd)